requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch
numpy
pyannote.audio
wavio
gradio
python-dotenv
typing-extensions
pytorch-lightning
torchmetrics
spaces